+++ lib/src/service/users/users_service.dart
@@ -265,6 +265,99 @@ abstract class UsersService {
   ///
   /// - https://developer.twitter.com/en/docs/twitter-api/users/lookup/api-reference/get-users-me
   Future<TwitterResponse<UserData, void>> lookupMe();
+
+  /// Allows an authenticated user ID to mute the target user.
+  ///
+  /// ## Parameters
+  ///
+  /// - [userId]: The user ID who you would like to initiate the mute on behalf
+  ///             of. It must match your own user ID or that of an
+  ///             authenticating user, meaning that you must pass the Access
+  ///             Tokens associated with the user ID when authenticating your
+  ///             request.
+  ///
+  /// - [targetUserId]: The user ID of the user that you would like the id to
+  ///                   mute. The body should contain a string of the user ID
+  ///                   inside of a JSON object.
+  ///
+  /// ## Endpoint Url
+  ///
+  /// - https://api.twitter.com/2/users/:id/muting
+  ///
+  /// ## Rate Limits
+  ///
+  /// - **User rate limit (OAuth 2.0 user Access Token)**:
+  ///    50 requests per 15-minute window per each authenticated user
+  ///
+  /// ## Reference
+  ///
+  /// - https://developer.twitter.com/en/docs/twitter-api/users/mutes/api-reference/post-users-user_id-muting
+  Future<bool> createMute(
+      {required String userId, required String targetUserId});
+
+  /// Allows an authenticated user ID to unmute the target user.
+  ///
+  /// The request succeeds with no action when the user sends a request to a
+  /// user they're not muting or have already unmuted.
+  ///
+  /// ## Parameters
+  ///
+  /// - [userId]: The user ID who you would like to initiate an unmute on
+  ///             behalf of. The user’s ID must correspond to the user ID of the
+  ///             authenticating user, meaning that you must pass the Access
+  ///             Tokens associated with the user ID when authenticating your
+  ///             request.
+  ///
+  /// - [targetUserId]: The user ID of the user that you would like the `userId`
+  ///                   to unmute.
+  ///
+  /// ## Endpoint Url
+  ///
+  /// - https://api.twitter.com/2/users/:source_user_id/muting/:target_user_id
+  ///
+  /// ## Rate Limits
+  ///
+  /// - **User rate limit (OAuth 2.0 user Access Token)**:
+  ///    50 requests per 15-minute window per each authenticated user
+  ///
+  /// ## Reference
+  ///
+  /// - https://developer.twitter.com/en/docs/twitter-api/users/mutes/api-reference/delete-users-user_id-muting
+  Future<bool> destroyMute(
+      {required String userId, required String targetUserId});
+
+  /// Returns a list of users who are muted by the specified user ID.
+  ///
+  /// ## Parameters
+  ///
+  /// - [userId]: The user ID whose muted users you would like to retrieve.
+  ///             The user’s ID must correspond to the user ID of the
+  ///             authenticating user, meaning that you must pass the Access
+  ///             Tokens associated with the user ID when authenticating your
+  ///             request.
+  ///
+  /// - [maxResults]: The maximum number of results to be returned per page.
+  ///                 This can be a number between 1 and 1000. By default, each
+  ///                 page will return 100 results.
+  ///
+  /// - [paginationToken]: Used to request the next page of results if all
+  ///                      results weren't returned with the latest request,
+  ///                      or to go back to the previous page of results.
+  ///
+  /// ## Endpoint Url
+  ///
+  /// - https://api.twitter.com/2/users/:id/muting
+  ///
+  /// ## Rate Limits
+  ///
+  /// - **User rate limit (OAuth 2.0 user Access Token)**:
+  ///    15 requests per 15-minute window per each authenticated user
+  ///
+  /// ## Reference
+  ///
+  /// - https://developer.twitter.com/en/docs/twitter-api/users/mutes/api-reference/get-users-muting
+  Future<TwitterResponse<List<UserData>, UserMeta>> mutingUsers(
+      {required String userId, int? maxResults, String? paginationToken});
 }
 
 class _UsersService extends BaseService implements UsersService {
@@ -406,5 +499,51 @@ class _UsersService extends BaseService implements UsersService {
     );
 
     return TwitterResponse(data: UserData.fromJson(response['data']));
+  }
+
+  @override
+  Future<bool> createMute(
+      {required String userId, required String targetUserId}) async {
+    final response = await super.post(
+      UserContext.oauth2OrOAuth1,
+      '/2/users/$userId/muting',
+      body: {'target_user_id': targetUserId},
+    );
+
+    return response['data']['muting'];
+  }
+
+  @override
+  Future<bool> destroyMute(
+      {required String userId, required String targetUserId}) async {
+    final response = await super.delete(
+      UserContext.oauth2OrOAuth1,
+      '/2/users/$userId/muting/$targetUserId',
+    );
+
+    return !response['data']['muting'];
+  }
+
+  @override
+  Future<TwitterResponse<List<UserData>, UserMeta>> mutingUsers({
+    required String userId,
+    int? maxResults,
+    String? paginationToken,
+  }) async {
+    final response = await super.get(
+      UserContext.oauth2OrOAuth1,
+      '/2/users/$userId/muting',
+      queryParameters: {
+        'max_results': maxResults,
+        'pagination_token': paginationToken,
+      },
+    );
+
+    return TwitterResponse(
+      data: response['data']
+          .map<UserData>((user) => UserData.fromJson(user))
+          .toList(),
+      meta: UserMeta.fromJson(response['meta']),
+    );
   }
 }